feat(callback): add request context
In addition to having the event emitter track the callbacks associated
with the token, also add the request arguments for a given token.

This allows us to pass that along with the callback. This gives the
client richer information about the context in which a callback was
emitted. This is useful for errors, where we can also display the
request context along with the actual error message.