fix(connector): add logic to close get requests
GET requests associated with a subscribe should be closed when the
subscribe is closed. The lack of this functionality caused a race
condition in the worker-based middleware when a subscription was closed
and re-opened before the associated GET request returned.

* Return SubscriptionIdentifier from callCommand.
* Return SubscriptionIdentifier instead of token from get and
getWithOptions.
* Create map from subscribe callbacks to GET SubscriptionIdentifier's
associated with them.
* Add logic to close GET's related to streams in closeStream and
closeStreams.
* Add logic to delete tokens from wrpc.activeRequests in
setStreamClosingState.
* Revise createCloseParams to remove repeated logic.
* Add handling to Connector.spec and Wrpc.spec for get/getWithOptions
returning a SubscriptionIdentifier instead of a simple token.