fix(msgpack): encode stripped neat objects
The encodeNeatClass function assumes that it receives an instantiated
Neat class (using instanceOf), but while it is always passed an object
that satisfies the constraints of a neat object, it may be passed a
neat object that has been stripped of its custom class (in this case,
passed across a worker socket).
Revise encodeNeatClass to use the `type` property istead of checking for
instanceOf.