fix(rollup): fix incorrect rollup configs
The Rollup config for cloudvision-connector references the `replace`
plugin, but this is the wrong package to do that, since that package
never uses `process.env` in its source code. This config belongs in
a-msgpack, which does access `process.env` in its source.

This fix will remove `process.env` from appearing in the build output
of a-msgpack, where it appears currently and causes non-webpack callers
to crash because `process` is not defined in a browser.